+++ tests/Unit/DSC_SqlAGReplica.Tests.ps1
@@ -1220,6 +1220,70 @@ Describe 'SqlAGReplica\Set-TargetResource' {
             }
         }
 
+        Context 'When multiple properties are not in desired state' {
+            BeforeAll {
+                Mock -CommandName Update-AvailabilityGroupReplica
+            }
+
+            It 'Should set all properties with one update' {
+                InModuleScope -Parameters $_ -ScriptBlock {
+                    $setTargetResourceParameters = @{
+                        Name                          = 'Server1'
+                        AvailabilityGroupName         = 'AG_AllServers'
+                        ServerName                    = 'Server1'
+                        InstanceName                  = 'MSSQLSERVER'
+                        PrimaryReplicaServerName      = 'Server2'
+                        PrimaryReplicaInstanceName    = 'MSSQLSERVER'
+                        Ensure                        = 'Present'
+                        AvailabilityMode              = 'SynchronousCommit'
+                        BackupPriority                = 60
+                        ConnectionModeInPrimaryRole   = 'AllowReadWriteConnections'
+                        ConnectionModeInSecondaryRole = 'AllowReadIntentConnectionsOnly'
+                        EndpointHostName              = 'AnotherEndpointHostName'
+                        FailoverMode                  = 'Automatic'
+                        ReadOnlyRoutingConnectionUrl  = 'TCP://Server1.domain.com:1433'
+                        ReadOnlyRoutingList           = @('Server2', 'Server1')
+                    }
+
+                    { Set-TargetResource @setTargetResourceParameters } | Should -Not -Throw
+                }
+
+                Should -Invoke -CommandName Update-AvailabilityGroupReplica -Exactly -Times 1 -Scope It
+            }
+        }
+
+        Context 'When AvailabilityMode and FailoverMode properties are not in desired state' {
+            BeforeAll {
+                Mock -CommandName Update-AvailabilityGroupReplica
+            }
+
+            It 'Should set both properties with one update' {
+                InModuleScope -Parameters $_ -ScriptBlock {
+                    $setTargetResourceParameters = @{
+                        Name                          = 'Server1'
+                        AvailabilityGroupName         = 'AG_AllServers'
+                        ServerName                    = 'Server1'
+                        InstanceName                  = 'MSSQLSERVER'
+                        PrimaryReplicaServerName      = 'Server2'
+                        PrimaryReplicaInstanceName    = 'MSSQLSERVER'
+                        Ensure                        = 'Present'
+                        AvailabilityMode              = 'SynchronousCommit'
+                        BackupPriority                = 50
+                        ConnectionModeInPrimaryRole   = 'AllowAllConnections'
+                        ConnectionModeInSecondaryRole = 'AllowNoConnections'
+                        EndpointHostName              = 'Server1'
+                        FailoverMode                  = 'Automatic'
+                        ReadOnlyRoutingConnectionUrl  = 'TCP://Server1.domain.com:1433'
+                        ReadOnlyRoutingList           = @('Server1', 'Server2')
+                    }
+
+                    { Set-TargetResource @setTargetResourceParameters } | Should -Not -Throw
+                }
+
+                Should -Invoke -CommandName Update-AvailabilityGroupReplica -Exactly -Times 1 -Scope It
+            }
+        }
+
         Context 'When the endpoint port differ from the port in the replica''s endpoint URL' {
             BeforeAll {
                 Mock -CommandName Update-AvailabilityGroupReplica